% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfd_pc_to_xy.R
\name{gen_face}
\alias{gen_face}
\title{generate a random face coordinate matrix from the pfd face space}
\usage{
gen_face(fixed_pcs = 0)
}
\arguments{
\item{the}{number of leading principal components to fix at zero.}
}
\value{
an 85 x 2 matrix of face coordinates
}
\description{
generate a random face coordinate matrix from the pfd face space
}
