% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfd_pc_to_xy.R
\name{gen_face_spline}
\alias{gen_face_spline}
\title{make a random face from parametric face space}
\usage{
gen_face_spline(fixed_pcs = 0)
}
\arguments{
\item{the}{number of leading principal components to fix at zero.}
}
\value{
an instance of class \code{face_spline}
}
\description{
generate a random face spline from the pfd face space
}
