% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfd_splines.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(x, nrow, ncol)
}
\arguments{
\item{x}{a list of \code{face_spline} objects}

\item{nrow}{the number of rows in plot layout}

\item{ncol}{the number of columns in plot layout}
}
\description{
Plot multiple faces in a grid layout
}
