% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfd_pc_to_xy.R
\name{pfd_pc_to_xy}
\alias{pfd_pc_to_xy}
\title{face pcs to xy coordinates}
\usage{
pfd_pc_to_xy(x)
}
\arguments{
\item{x}{a face represented by an 85 x 2 matrix}
}
\value{
an 85 x 2 matrix of face coordinates
}
\description{
convert face pc coefficients to xy face coordinates
}
