% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfd_splines.R
\name{pfd_splines}
\alias{pfd_splines}
\title{convert face coordinates to a drawable shape}
\usage{
pfd_splines(face, spline_shape = 1)
}
\arguments{
\item{face}{an 85 x 2 matrix of face coordinates}

\item{spline_shape}{1 for cubic spline, -1 for Catmull-Rom splines, 0 for polygons}
}
\description{
convert face coordinates to a drawable shape
}
